+++ packages/radio/tsconfig.spec.json
@@ -15,6 +15,7 @@
     "src/**/*.spec.js",
     "src/**/*.test.jsx",
     "src/**/*.spec.jsx",
-    "src/**/*.d.ts"
+    "src/**/*.d.ts",
+    "*/.stories.tsx"
   ]
 }
